relay-client/app/requirements.txt
paho.mqtt
Homie4